requirements.txt
pandas
numpy
scikit-learn
mlflow
scipy
dvc 
pyarrow 
hydra-core
flatten-dict
prefect o
pytest
flake8 